requirements.txt
fastapi
torchaudio
torchvision
GPUtil
